FEAT(3iD): Service extraction
