+++ 3iD/src/services/threeid/index.ts
@@ -1,0 +1,17 @@
+import { getFunnelState, tickFunnelStep } from "./funnel";
+import { claimInvitation, listInvitations } from "./inviting";
+import { genPfP } from "./minting";
+import { getENSEntries } from "./naming";
+import { fetchProfile, fetchPublicProfile, setProfile } from "./profile";
+
+export {
+  fetchPublicProfile,
+  getFunnelState,
+  listInvitations,
+  tickFunnelStep,
+  claimInvitation,
+  genPfP,
+  fetchProfile,
+  setProfile,
+  getENSEntries,
+};
