--- 3iD/src/types/Invitation.d.ts
@@ -1,10 +1,0 @@
-export interface Invitation {
-  contractAddress: string;
-  tokenId: string;
-  title: string;
-
-  /**
-   * URL of the image
-   */
-  image: string;
-}
