--- 3iD/src/types/Profile.d.ts
@@ -1,20 +1,0 @@
-export interface Profile {
-  nickname?: string;
-  profilePicture?: {
-    collectionTokenId: string;
-    collectionId?: string;
-    name: string;
-    imageUrl: string;
-  };
-  email?: string;
-  location?: string;
-  job?: string;
-  website?: string;
-  bio?: string;
-  socials?: {
-    twitter?: string;
-    instagram?: string;
-    linkedin?: string;
-    github?: string;
-  };
-}
